Fixed the weird flicker I was getting sometimes by making sortInPlace() accept comparison functions that return any type that can be compared with 0. But, by chance I found a bug in the debug system that I don't understand - the linked lists get corrupted at some point, but it's hard to know where or when because I've never seen it before, and it's #define macro code so I can't inspect it properly. Should probably rip that out and replace it with something less awkward. CLOC: 9557

+++ src/render.cpp
@@ -125,10 +125,9 @@ void drawRenderItem(RenderBuffer *buffer, RenderItem *item, V2 offsetP, f32 dept
 	dest->textureRegionID = item->textureRegionID;
 }
 
-s32 compareRenderItems(RenderItem *a, RenderItem *b)
+f32 compareRenderItems(RenderItem *a, RenderItem *b)
 {
-	// NB: Hmmm. Having to convert float -> int might be a problem.
-	return (s32)(a->depth - b->depth);
+	return (a->depth - b->depth);
 }
 
 void sortRenderBuffer(RenderBuffer *buffer)
